requirements.txt
Flask==1.1.2
gunicorn==20.0.4
matplotlib==3.3.1
numpy==1.19.1
opencv-python==4.4.0.44
pandas==1.1.1
scikit-learn==0.23.2
sklearn==0.0
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.6
os
glob
time
cv2.cv2